requirements.txt
openai==1.3.0
httpx==0.24.1
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
flask==2.3.3
python-dotenv==1.0.0
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0 